rename place_log_util to placement_log_printer
