+++ vpr/src/place/placement_log_printer.cpp
@@ -1,5 +1,5 @@
 
-#include "place_log_util.h"
+#include "placement_log_printer.h"
 
 #include "vtr_log.h"
 #include "annealer.h"
